requirements.txt
Django==1.11
django-cors-headers==2.1.0
django-taggit==0.22.2
djangorestframework==3.7.7
gunicorn==19.7.1
pytz==2017.3
whitenoise==3.3.1
